==> data/00_scripts/10_data-wrangling.Rmd <==
---
title: "10_data-wrangling"
author: "Alma Garcia"
date: "2024-03-02"
output: html_document
---

# Data Wrangling

## Metadata
### file import

```{r}
library(tidyverse)
library(readxl)
```

```{r}
human_metadata_raw <- read_xlsx("../10_metadata-and-manifest/original_files/original_human_metadata.xlsx")
mouse_metadata_raw <- read.delim("../10_metadata-and-manifest/original_files/original_mouse_metadata.tsv")
```

### filtering

We will keep only the data from timepoint 2 and 4 
```{r}
human_metadata <- human_metadata_raw %>% 
									filter(timepoint %in% c(2,4)) %>%
									select("#SampleID",host_subject_id, timepoint, trt, trtseq, gender) %>%
									rename(`sample-id` = "#SampleID")
```

We will remove the positive control INU samples because they're not a resistant starch treatment and can't be compared between the datasets. 
```{r}
mouse_metadata <- mouse_metadata_raw %>% 
									filter(Treatment != "INU") %>%
									# add treatment type column to match the human dataset
									mutate(trt = ifelse(Treatment == "CTL", "C", "RS"),
									# add gender column
												 gender = ifelse(str_detect(sample.id,"F"), "F", "M")
												 ) %>%
									# rename for QIIME
									rename(`sample-id` = sample.id)
```

### merge
Add cohort columns
```{r}
human_metadata <- human_metadata %>% mutate(cohort = "human")
mouse_metadata <- mouse_metadata %>% mutate(cohort = "mouse")
```

Merge tables and cleanup:
```{r}
metadata <- full_join(human_metadata, mouse_metadata) %>%
	# drop unused columns
	select(-c(timepoint, sample_type, host_subject_id, strain)) %>%
	# rename Treatment to starch_type
	rename(starch_type = Treatment) %>%
	# create treatment group column (organism-treatment)
	mutate(group = str_c(cohort, trt, sep = "-"))
```

### export

```{r}
write_tsv(human_metadata, "../10_metadata-and-manifest/human_metadata.tsv")
write_tsv(mouse_metadata, "../10_metadata-and-manifest/mouse_metadata.tsv")
write_tsv(metadata, "../10_metadata-and-manifest/metadata.tsv")
```

## Manifest`

Import old manifest files:
```{r}
mouse_manifest_old <- read.delim("../10_metadata-and-manifest/original_files/original_mouse_manifest.tsv")
human_manifest_old <- read.delim("../10_metadata-and-manifest/original_files/original_human_manifest.txt")
```

Filter manifest table to only include samples in metadata files:
```{r}
mouse_manifest <- mouse_manifest_old %>% filter(sample.id %in% mouse_metadata$`sample-id`)
human_manifest <- human_manifest_old %>% filter(sample.id %in% human_metadata$`sample-id`)
```

Correct mouse data path:
```{r}
mouse_manifest$absolute.filepath <- str_remove_all(mouse_manifest$absolute.filepath, "diabetes/")
```

Export new manifest files:
```{r}
write_tsv(mouse_manifest, "../10_metadata-and-manifest/mouse_manifest.tsv")
write_tsv(human_manifest, "../10_metadata-and-manifest/human_manifest.tsv")
```
