==> data/00_scripts/22_filtering-rarefying.Rmd <==
---
title: "Filter and Rarefying"
author: "Alma Garcia"
date: "2024-03-04"
output: html_document
---

# Filtering and Rarefying
Import libraries and data
```{r}
library(phyloseq)
library(vegan)
library(tidyverse)
```

```{r}
starch_unfiltered <- readRDS("../21_phyloseq/starch_unfiltered_phyloseq.RDS")
```

## Filtering
Removing low read count samples:
```{r}
sort(sample_sums(starch_unfiltered))
```

```{r}
# add total number of reads
starch_unfiltered@sam_data$sample_sums <- as.numeric(sample_sums(starch_unfiltered))

# prune by sample_sums
starch_high <- prune_samples(sample_sums(starch_unfiltered) >= 300, starch_unfiltered)
```


Save filtered non-rarefied phyloseq
```{r}
write_rds(starch_high, "../21_phyloseq/starch_filtered_phyloseq.RDS")
```


## Rarefying
```{r}
depth <- 6747
```

```{r}
otu_matrix <- as.matrix(as.data.frame(otu_table(starch_high)))
rare <- rarecurve(as.matrix(otu_matrix), step=500, label = FALSE, tidy = TRUE)
```

```{r}
rare %>%
	ggplot(aes(x = Sample, y = Species, group = Site)) +
	geom_line() +
	geom_vline(xintercept = depth, color = "red") +
	scale_x_continuous(labels = scales::comma) +
	coord_cartesian(xlim=c(0,10000)) +
	theme_bw() +
	theme(legend.position = "none") 
```

```{r}
dephyloseq <- function(phylo_obj){ 
	## get the metadata
	meta = as.data.frame(as.matrix(phylo_obj@sam_data)) 
	## how many metadata columns you have
	metacols = ncol(meta)+1
	## get out the otu table
	## if your metadta is empty after running this, you need to use #otu = as.data.frame(t(as.matrix(phylo_obj@otu_table)))
	otu = as.data.frame(t(as.matrix(phylo_obj@otu_table)))
	## merge the metadata and otu table by the rownames (sample ids from the Illumina sequence)
	mo = merge(meta, otu, by=0) 
	## get out the taxonomy file
	tax = as.data.frame(phylo_obj@tax_table)
	## get the ASV ID out. This the matches the placeholder ASV ID in the OTU table
	tax = tax %>% rownames_to_column(var="asv_id")
	## pivot longer to be able to match the ASVs in the OTU table to the taxonomy table
	mo = mo %>% pivot_longer(cols = -c(1:metacols), names_to = "asv_id", values_to="asv_abundance") 
	## Join the metadata and otu table with the taoxnomy table
	mot = full_join(mo, tax)
  ## Specify the output for the dephyloseq funciton
output = mot 
}
```


```{r}
starch_df <- dephyloseq(starch_high)
```

```{r}
df <- starch_df %>%
	select(c(Row.names, group, sample_sums, gender)) %>%
	rename(`sample-id` = Row.names) %>%
	distinct()
```

```{r}
df$sample_sums <- as.numeric(df$sample_sums)
df$color <- if_else(df$sample_sums>=depth, "keep", "discard")
```

```{r}
df %>% ggplot(aes(as.character(`sample-id`), as.numeric(sample_sums))) + 
	geom_col(aes(fill = color)) + 
	geom_hline(aes(yintercept = depth), color = "red") +
	facet_wrap(.~group,
						 scales = "free") +
	scale_fill_manual(values = c("grey", "black")) +
	ylab("sample sums") +
	theme_minimal() +
	theme(
		axis.text.x = element_text(angle = 45, hjust = 1, vjust = 1),
		legend.position = "none",
		axis.title.x = element_blank()
	)
```

```{r}
starch_rare <- rarefy_even_depth(starch_taxa, rngseed = 1, sample.size = depth)
write_rds(starch_rare, "../21_phyloseq/starch_rarefied_phyloseq.RDS")
```
