==> data/00_scripts/30_diversity-analysis.Rmd <==
---
title: "Diversity Analysis"
output:
  html_document:
    toc: true
    toc_depth: 4
    toc_float:
      collapsed: false
      smooth_scroll: true
---

# Diversity Analysis
Import libraries and data
```{r}
library(phyloseq)
library(ape)
library(tidyverse)
library(picante)
library(pairwiseAdonis)
library(ggpubr)
```

```{r}
starch <- read_rds("../21_phyloseq/starch_rarefied_phyloseq.RDS")
```

Separate human and mouse for the independent analysis
```{r}
human <- subset_samples(starch, cohort == "human")
mouse <- subset_samples(starch, cohort == "mouse")
```

```{r}
pal <- c("#DD4124FF", "#faa478", "#0269a1", "#6ac1de")
pal2 <- c("#00346EFF", "#007CBFFF", "#06ABDFFF", "#EDD03EFF", "#F5A200FF", "#6D8600FF", "#424D0CFF")
pal3 <- c("#FDAE61FF", "#ABDDA4FF")
```


## Alpha diversity

```{r}
richness <- estimate_richness(starch)
rownames(richness) <- rownames(starch@sam_data)
```

```{r}
starch_rich <- merge(starch@sam_data, richness, by = 0) %>% dplyr::rename(sample_id = Row.names)
```

Reordering group factors so that they make sense
```{r}
starch_rich$group <- factor(starch_rich$group, 
														levels = c("human-C", "human-RS", 
																			 "mouse-C", "mouse-RS"))

```


### Shannon and Chao1
#### Statistical Analysis
##### Merged
```{r}
shannon_kw <- kruskal.test(Shannon ~ group, data = starch_rich)
shannon_kw

if (shannon_kw$p.value < 0.05){
	pairwise.wilcox.test(starch_rich$Shannon, starch_rich$group, 
                       p.adjust = "bonferroni")
}
```
There is a significant difference (p-val < 0.05) in the Shannon index between: 
- mouse-C & human-C
- mouse-C & human-RS
- mouse-RS & human-C

```{r}
shannon_kw <- kruskal.test(Shannon ~ gender, data = starch_rich)
shannon_kw
```
Overall shows a significant difference in the Shannon index between genders.

```{r}
chao1_kw <- kruskal.test(Chao1 ~ group, data = starch_rich)
chao1_kw

if (chao1_kw$p.value < 0.05){
	pairwise.wilcox.test(starch_rich$Chao1, starch_rich$group, 
                       p.adjust = "bonferroni")
}
```
Difference in Chao1 index between: human-c & mouse-c and human-C & mouse-rs
No difference between: human-RS and mouse-RS!!

```{r}
chao1_kw <- kruskal.test(Chao1 ~ gender, data = starch_rich)
chao1_kw

if (chao1_kw$p.value < 0.05){
	pairwise.wilcox.test(starch_rich$Chao1, starch_rich$gender, 
                       p.adjust = "bonferroni")
}
```
No significant difference between genders. 

##### Mouse

```{r}
starch_rich %>% 
	filter(cohort == "mouse") %>%
	kruskal.test(Shannon ~ trt, data = .)
```
No significant difference in Shannon index between treatments.

```{r}
starch_rich %>% 
	filter(cohort == "mouse") %>%
	kruskal.test(Chao1 ~ trt, data = .)
```
No significant difference in Chao1 index between treatments.


```{r}
starch_rich %>% 
	filter(cohort == "mouse") %>%
	kruskal.test(Shannon ~ gender, data = .)
```

Significant difference in Shannon index between genders.

```{r}
starch_rich %>% 
	filter(cohort == "mouse") %>%
	kruskal.test(Chao1 ~ gender, data = .)
```
Significant difference in Chao1 index between genders.

```{r}
starch_rich %>% 
	filter(cohort == "mouse") %>%
	kruskal.test(Shannon ~ starch_type, data = .)
```
No significant difference in Shannon index between starch types

```{r}
mouse_kw <- starch_rich %>% 
	filter(cohort == "mouse") %>%
	kruskal.test(Chao1 ~ starch_type, data = .)
```

```{r}
mouse_rich <- starch_rich %>% 
	filter(cohort == "mouse")

chao1_kw <- mouse_rich %>%
	kruskal.test(Chao1 ~ starch_type, data = .)

if (chao1_kw$p.value < 0.05){
	pairwise.wilcox.test(mouse_rich$Chao1, mouse_rich$starch_type, 
                       p.adjust = "bonferroni")
}
```

Significant difference in mean Chao1 index between CKP and PTB. 

##### Human
```{r}
starch_rich %>% 
	filter(cohort == "human") %>%
	kruskal.test(Shannon ~ trt, data = .)
```
Significant difference in Shannon index between treatments 

```{r}
starch_rich %>% 
	filter(cohort == "human") %>%
	kruskal.test(Chao1 ~ trt, data = .)
```
No significant difference in Chao1 index between treatments. 

```{r}
starch_rich %>% 
	filter(cohort == "human") %>%
	kruskal.test(Shannon ~ gender, data = .)
```
Significant difference in Shannon index between genders. 

```{r}
starch_rich %>% 
	filter(cohort == "human") %>%
	kruskal.test(Chao1 ~ gender, data = .)
```
No significant difference in Chao1 index between genders. 


#### Plots

```{r}
my_comparisons <- list(c("human-C", "human-RS"),
											 c("mouse-C", "mouse-RS"),
											 c("mouse-RS","human-RS"),
											 c("mouse-C","human-C"))
starch_comparisons <- list(c("CTL", "CKP"),
													 c("CKP", "LEN"),
													 c("CKP", "BEP"),
													 c("CKP", "PTB")
													 )
```


##### Merged

```{r}
w <- 1800
h <- 1500
font_size <- 18
```


```{r}
starch_rich %>%
	ggboxplot(x = "group", y = "Shannon", color =  "group",
						palette = pal) +
	stat_compare_means(comparisons = my_comparisons,
										 method = "wilcox.test",
										 hide.ns = FALSE,
										 label = "p.signif") +
	theme_bw(base_size = font_size) +
	theme(
		panel.grid.major = element_blank(),
		panel.grid.minor = element_blank(),
		legend.position = "none",
		axis.title.x = element_blank()
	)

ggsave("../30_diversity-analysis/plots/alpha-diversity/shannon_group.png", units = "px", width = w, height = h, create.dir = TRUE)

starch_rich %>%
	ggboxplot(x = "gender", y = "Shannon", color =  "gender",
						palette = pal3) +
	stat_compare_means(method = "kruskal.test", 
										 hide.ns = TRUE,
										 label = "p.signif") +
	theme_bw(base_size = font_size) +
	theme(
		panel.grid.major = element_blank(),
		panel.grid.minor = element_blank(),
		legend.position = "none",
		axis.title.x = element_blank()
	)

ggsave("../30_diversity-analysis/plots/alpha-diversity/shannon_gender.png", units = "px", width = w, height = h, create.dir = TRUE)

starch_rich %>%
	ggboxplot(x = "gender", y = "Shannon", color =  "gender",
						palette = pal3, facet.by = "group") +
	stat_compare_means(method = "kruskal.test", 
										 hide.ns = TRUE,
										 label = "p.signif") +
	theme_bw(base_size = font_size) +
	theme(
		panel.grid.major = element_blank(),
		panel.grid.minor = element_blank(),
		legend.position = "none",
		axis.title.x = element_blank()
	)

ggsave("../30_diversity-analysis/plots/alpha-diversity/shannon_gender_facet.png", units = "px", width = w, height = h, create.dir = TRUE)
```

```{r}
starch_rich %>%
	ggboxplot(x = "group", y = "Chao1", color =  "group",
						palette = pal) +
	stat_compare_means(comparisons = my_comparisons,
										 method = "wilcox.test", 
										 hide.ns = FALSE,
										 label = "p.signif") +
	theme_bw(base_size = font_size) +
	theme(
		panel.grid.major = element_blank(),
		panel.grid.minor = element_blank(),
		legend.position = "none",
		axis.title.x = element_blank()
	)

ggsave("../30_diversity-analysis/plots/alpha-diversity/chao1_group.png", units = "px", width = w, height = h, create.dir = TRUE)

starch_rich %>%
	ggboxplot(x = "gender", y = "Chao1", color =  "gender",
						palette = pal3) +
	stat_compare_means(method = "kruskal.test", 
										 hide.ns = TRUE,
										 label = "p.signif") +
	theme_bw(base_size = font_size) +
	theme(
		panel.grid.major = element_blank(),
		panel.grid.minor = element_blank(),
		legend.position = "none",
		axis.title.x = element_blank()
	)

ggsave("../30_diversity-analysis/plots/alpha-diversity/chao1_gender.png", units = "px", width = w, height = h, create.dir = TRUE)

starch_rich %>%
	ggboxplot(x = "gender", y = "Chao1", color = "gender",
						palette = pal3, facet.by = "group") +
	stat_compare_means(method = "kruskal.test", 
										 hide.ns = TRUE,
										 label = "p.signif") +
	theme_bw(base_size = font_size) +
	theme(
		panel.grid.major = element_blank(),
		panel.grid.minor = element_blank(),
		legend.position = "none",
		axis.title.x = element_blank()
	)

ggsave("../30_diversity-analysis/plots/alpha-diversity/chao1_gender_facet.png", units = "px", width = w, height = h, create.dir = TRUE)
```

##### Mouse
```{r}
starch_rich %>%
	filter(cohort == "mouse") %>%
	ggboxplot(x = "gender", y = "Shannon", color =  "gender",
						palette = pal3) +
	facet_wrap(.~trt) +
	stat_compare_means(method = "kruskal.test",
										 hide.ns = TRUE,
										 label = "p.signif") +
	theme_bw() +
	theme(
		panel.grid.major = element_blank(),
		panel.grid.minor = element_blank(),
		legend.position = "right"
	)

ggsave("../30_diversity-analysis/plots/alpha-diversity/shannon_gender_mouse.png", units = "px", width = w, height = h, create.dir = TRUE)

starch_rich %>%
	filter(cohort == "mouse") %>%
	ggboxplot(x = "trt", y = "Shannon", color =  "trt",
						palette = rev(pal[3:4])) +
	facet_wrap(.~gender) +
	stat_compare_means(method = "kruskal.test",
										 hide.ns = TRUE,
										 label = "p.signif") +
	theme_bw() +
	theme(
		panel.grid.major = element_blank(),
		panel.grid.minor = element_blank(),
		legend.position = "right"
	)

ggsave("../30_diversity-analysis/plots/alpha-diversity/shannon_trt_mouse.png", units = "px", width = w, height = h, create.dir = TRUE)

starch_rich %>%
	filter(cohort == "mouse") %>%
	ggboxplot(x = "starch_type", y = "Shannon", color =  "starch_type",
						palette = pal2) +
	stat_compare_means(comparisons = starch_comparisons,
										 method = "kruskal.test",
										 hide.ns = TRUE,
										 label = "p.signif") +
	xlab("Starch Type") +
	theme_bw() +
	theme(
		panel.grid.major = element_blank(),
		panel.grid.minor = element_blank(),
		legend.position = "none"
	)

ggsave("../30_diversity-analysis/plots/alpha-diversity/shannon_type_mouse.png", units = "px", width = w, height = h, create.dir = TRUE)
```

```{r}
starch_rich %>%
	filter(cohort == "mouse") %>%
	ggboxplot(x = "gender", y = "Chao1", color =  "gender",
						palette = pal3) +
	facet_wrap(.~trt) +
	stat_compare_means(method = "kruskal.test",
										 hide.ns = TRUE,
										 label = "p.signif") +
	theme_bw() +
	theme(
		panel.grid.major = element_blank(),
		panel.grid.minor = element_blank(),
		legend.position = "right"
	)

ggsave("../30_diversity-analysis/plots/alpha-diversity/chao1_gender_mouse.png", units = "px", width = w, height = h, create.dir = TRUE)

starch_rich %>%
	filter(cohort == "mouse") %>%
	ggboxplot(x = "trt", y = "Chao1", color =  "trt",
						palette = rev(pal[3:4])) +
	facet_wrap(.~gender) +
	stat_compare_means(method = "kruskal.test",
										 hide.ns = TRUE,
										 label = "p.signif") +
	theme_bw() +
	theme(
		panel.grid.major = element_blank(),
		panel.grid.minor = element_blank(),
		legend.position = "right"
	)

ggsave("../30_diversity-analysis/plots/alpha-diversity/chao1_trt_mouse.png", units = "px", width = w, height = h, create.dir = TRUE)

starch_rich %>%
	filter(cohort == "mouse") %>%
	ggboxplot(x = "starch_type", y = "Chao1", color =  "starch_type",
						palette = pal2) +
	stat_compare_means(comparisons = starch_comparisons,
										 method = "kruskal.test",
										 hide.ns = TRUE,
										 label = "p.signif") +
	xlab("Starch Type") +
	theme_bw() +
	theme(
		panel.grid.major = element_blank(),
		panel.grid.minor = element_blank(),
		legend.position = "none"
	)

ggsave("../30_diversity-analysis/plots/alpha-diversity/chao1_type_mouse.png", units = "px", width = w, height = h, create.dir = TRUE)
```

##### Human
```{r}
starch_rich %>%
	filter(cohort == "human") %>%
	ggboxplot(x = "gender", y = "Shannon", color =  "gender",
						palette = pal3) +
	facet_wrap(.~trt) +
	stat_compare_means(method = "kruskal.test",
										 hide.ns = TRUE,
										 label = "p.signif") +
	theme_bw() +
	theme(
		panel.grid.major = element_blank(),
		panel.grid.minor = element_blank(),
		legend.position = "right"
	)

ggsave("../30_diversity-analysis/plots/alpha-diversity/shannon_gender_human.png", units = "px", width = w, height = h, create.dir = TRUE)

starch_rich %>%
	filter(cohort == "human") %>%
	ggboxplot(x = "trt", y = "Shannon", color =  "trt",
						palette = pal) +
	facet_wrap(.~gender) +
	stat_compare_means(method = "kruskal.test", 
										 hide.ns = TRUE,
										 label = "p.signif") +
	theme_bw() +
	theme(
		panel.grid.major = element_blank(),
		panel.grid.minor = element_blank(),
		legend.position = "right"
	)

ggsave("../30_diversity-analysis/plots/alpha-diversity/shannon_trt_human.png", units = "px", width = w, height = h, create.dir = TRUE)
```

```{r}
starch_rich %>%
	filter(cohort == "human") %>%
	ggboxplot(x = "gender", y = "Chao1", color =  "gender",
						palette = pal3) +
	facet_wrap(.~trt) +
	stat_compare_means(method = "kruskal.test",
										 hide.ns = TRUE,
										 label = "p.signif") +
	theme_bw() +
	theme(
		panel.grid.major = element_blank(),
		panel.grid.minor = element_blank(),
		legend.position = "right"
	)

ggsave("../30_diversity-analysis/plots/alpha-diversity/chao1_gender_human.png", units = "px", width = w, height = h, create.dir = TRUE)

starch_rich %>%
	filter(cohort == "human") %>%
	ggboxplot(x = "trt", y = "Chao1", color =  "trt",
						palette = pal) +
	facet_wrap(.~gender) +
	stat_compare_means(method = "kruskal.test", 
										 hide.ns = TRUE,
										 label = "p.signif") +
	theme_bw() +
	theme(
		panel.grid.major = element_blank(),
		panel.grid.minor = element_blank(),
		legend.position = "right"
	)

ggsave("../30_diversity-analysis/plots/alpha-diversity/chao1_trt_human.png", units = "px", width = w, height = h, create.dir = TRUE)
```


### Faith's Phylogenetic Diversity 
```{r}
phylo_dist <- pd(t(otu_table(starch)), phy_tree(starch),
                 include.root=F) 

# add PD to metadata table
starch_rich$PD <- phylo_dist$PD
```

#### Statistical Analysis
##### Group
```{r}
pd_kw <- kruskal.test(PD ~ group, data = starch_rich)
pd_kw

if (pd_kw$p.value < 0.05){
	pairwise.wilcox.test(starch_rich$PD, starch_rich$group, 
                       p.adjust = "bonferroni")
}
```
**There is no significant difference in Faith's phylogenetic distance between the groups.**

##### Gender
```{r}
kruskal.test(PD ~ gender, data = starch_rich)
```
No difference between genders. 

##### Mouse
```{r}
starch_rich %>% 
	filter(cohort == "mouse") %>%
	kruskal.test(PD ~ gender, data = .)
```
No difference between genders.

```{r}
starch_rich %>% 
	filter(cohort == "mouse") %>%
	kruskal.test(PD ~ starch_type, data = .)
```
No difference between starch types. 

##### Human
```{r}
starch_rich %>% 
	filter(cohort == "human") %>%
	kruskal.test(PD ~ gender, data = .)
```
No difference between genders. 

#### Plots
##### Merged
```{r}
starch_rich %>%
	ggboxplot(x = "group", y = "PD", color =  "group",
						palette = pal) +
	theme_bw(base_size = font_size) +
	ylab("Faith's phylogenetic diversity") +
	theme(
		panel.grid.major = element_blank(),
		panel.grid.minor = element_blank(),
		legend.position = "none",
		axis.title.x = element_blank()
	)

ggsave("../30_diversity-analysis/plots/alpha-diversity/pd_group.png", units = "px", width = w, height = h, create.dir = TRUE)

starch_rich %>%
	ggboxplot(x = "gender", y = "PD", color =  "gender",
						palette = pal3) +
	theme_bw(base_size = font_size) +
	ylab("Faith's phylogenetic diversity") +
	theme(
		panel.grid.major = element_blank(),
		panel.grid.minor = element_blank(),
		legend.position = "none"
	)

ggsave("../30_diversity-analysis/plots/alpha-diversity/pd_gender.png", units = "px", width = w, height = h, create.dir = TRUE)

starch_rich %>%
	ggboxplot(x = "gender", y = "PD", color =  "gender",
						palette = pal3, facet.by = "group") +
	theme_bw(base_size = font_size) +
	ylab("Faith's phylogenetic diversity") +
	theme(
		panel.grid.major = element_blank(),
		panel.grid.minor = element_blank(),
		legend.position = "none"
	)

ggsave("../30_diversity-analysis/plots/alpha-diversity/pd_gender_facet.png", units = "px", width = w, height = h, create.dir = TRUE)
```

## Beta Diversity
```{r}
starchdf <- as.data.frame(sample_data(starch))
humandf <- as.data.frame(sample_data(human))
mousedf <- as.data.frame(sample_data(mouse))
```

Reordering group levels


```{r}
starchdf$group <- factor(starchdf$group, 
														levels = c("human-C", "human-RS", 
																			 "mouse-C", "mouse-RS"))
```

```{r}
font_size <- 14
```


### Bray Curtis

#### Merged
```{r}
bc_dm <- phyloseq::distance(starch, method="bray")
adonis2(bc_dm ~ group, data = data.frame(starchdf))
adonis2(bc_dm ~ group*gender, data = data.frame(starchdf))
adonis2(bc_dm ~ trt*cohort*gender, data = data.frame(starchdf))
```
There is a significant difference between treatments, cohorts, genders, at least two trt:gender pair, and at least two cohort:gender pair.

```{r}
pcoa_bc <- ordinate(starch, method="PCoA", distance=bc_dm)

plot_ordination(starch, pcoa_bc, color = "group") +
	scale_color_manual(values = pal) +
  labs(col = "Group") +
	theme_bw(base_size = font_size) +
	theme(
		panel.grid.major = element_blank(),
		panel.grid.minor = element_blank(),
	)

ggsave("../30_diversity-analysis/plots/beta-diversity/bc_group.png", units = "px", width = w, height = h, create.dir = TRUE)

plot_ordination(starch, pcoa_bc, color = "gender") +
	scale_color_manual(values = pal3) +
  labs(col = "Gender") +
	theme_bw(base_size = font_size) +
	theme(
		panel.grid.major = element_blank(),
		panel.grid.minor = element_blank(),
	)

ggsave("../30_diversity-analysis/plots/beta-diversity/bc_gender.png", units = "px", width = w, height = h, create.dir = TRUE)
```
#### Human
```{r}
human_bc_dm <- phyloseq::distance(human, method="bray")
adonis2(human_bc_dm ~ trt*gender, data = data.frame(humandf))
```
There is only a significant difference in community structure between genders. I should probably consider here the effect of `trtseq` (fixed effect) or individuals as random error....

```{r}
human_pcoa_bc <- ordinate(human, method="PCoA", distance=human_bc_dm)

plot_ordination(human, human_pcoa_bc, color = "trt") +
	scale_color_manual(values = pal) +
  labs(col = "Treatment") +
	theme_bw(base_size = font_size) +
	theme(
		panel.grid.major = element_blank(),
		panel.grid.minor = element_blank(),
	)

ggsave("../30_diversity-analysis/plots/beta-diversity/bc_trt_human.png", units = "px", width = w, height = h, create.dir = TRUE)

plot_ordination(human, human_pcoa_bc, color = "gender") +
	scale_color_manual(values = pal3) +
  labs(col = "Gender") +
	theme_bw(base_size = font_size) +
	theme(
		panel.grid.major = element_blank(),
		panel.grid.minor = element_blank(),
	)

ggsave("../30_diversity-analysis/plots/beta-diversity/bc_gender_human.png", units = "px", width = w, height = h, create.dir = TRUE)

plot_ordination(human, human_pcoa_bc, color = "gender") +
	scale_color_manual(values = pal3) +
	facet_grid(.~trt) +
  labs(col = "Gender") +
	theme_bw(base_size = font_size) +
	theme(
		panel.grid.major = element_blank(),
		panel.grid.minor = element_blank(),
	)

ggsave("../30_diversity-analysis/plots/beta-diversity/bc_gender_facet_human.png", units = "px", width = w, height = h, create.dir = TRUE)
```

#### Mouse
```{r}
mouse_bc_dm <- phyloseq::distance(mouse, method="bray")
adonis2(mouse_bc_dm ~ trt*starch_type*gender, data = data.frame(mousedf))
pw <- pairwiseAdonis::pairwise.adonis2(mouse_bc_dm ~ trt*starch_type*gender, 
																			 data = data.frame(mousedf))

print("significantly different pairs: ")
for (i in range(2,(length(pw)))){
	pair <- pw[[i]]
	if (pair$`Pr(>F)`[1] < 0.05) {
		print(names(pw)[i])
	}
}
```
There is a significant difference in community structure between treatments, at least 2 starch types, genders, at least two trt:gender pairs, and at least two starch_type:gender pairs. 

```{r}
mouse_pcoa_bc <- ordinate(mouse, method="PCoA", distance=mouse_bc_dm)

plot_ordination(mouse, mouse_pcoa_bc, color = "trt") +
	#stat_ellipse(alpha = 0.5) +
	scale_color_manual(values = pal[3:4]) +
  labs(col = "Treatment") +
	theme_bw(base_size = font_size) +
	theme(
		panel.grid.major = element_blank(),
		panel.grid.minor = element_blank(),
	)

ggsave("../30_diversity-analysis/plots/beta-diversity/bc_trt_mouse.png", units = "px", width = w, height = h, create.dir = TRUE)

plot_ordination(mouse, mouse_pcoa_bc, color = "starch_type") +
	#stat_ellipse(alpha = 0.5) +
	scale_color_manual(values = pal2) +
  labs(col = "Starch Type") +
	theme_bw(base_size = font_size) +
	theme(
		panel.grid.major = element_blank(),
		panel.grid.minor = element_blank(),
	)

ggsave("../30_diversity-analysis/plots/beta-diversity/bc_type_mouse.png", units = "px", width = w, height = h, create.dir = TRUE)

plot_ordination(mouse, mouse_pcoa_bc, color = "gender") +
	#stat_ellipse(alpha = 0.5) +
	scale_color_manual(values = pal3) +
  labs(col = "Starch Type") +
	theme_bw(base_size = font_size) +
	theme(
		panel.grid.major = element_blank(),
		panel.grid.minor = element_blank(),
	)

ggsave("../30_diversity-analysis/plots/beta-diversity/bc_gender_mouse.png", units = "px", width = w, height = h, create.dir = TRUE)
```
### Jaccard
#### Merged
```{r}
jc_dm <- phyloseq::distance(starch, method = "jaccard", binary = TRUE)
adonis2(jc_dm ~ group*gender, data = data.frame(starchdf))
adonis2(jc_dm ~ cohort*trt*gender, data = data.frame(starchdf))
```
There is a significant difference in community composition between cohorts, genders, and at least 2 cohort:gender pairs. 

```{r}
pcoa_jc <- ordinate(starch, method="PCoA", distance=jc_dm)

plot_ordination(starch, pcoa_jc, color = "group") +
	#stat_ellipse(alpha = 0.5) +
	scale_color_manual(values = pal) +
  labs(col = "Group") +
	theme_bw(base_size = font_size) +
	theme(
		panel.grid.major = element_blank(),
		panel.grid.minor = element_blank(),
	)

ggsave("../30_diversity-analysis/plots/beta-diversity/jc_group.png", units = "px", width = w, height = h, create.dir = TRUE)

plot_ordination(starch, pcoa_jc, color = "gender") +
	#stat_ellipse(alpha = 0.5) +
	scale_color_manual(values = pal3) +
  labs(col = "Gender") +
	theme_bw(base_size = font_size) +
	theme(
		panel.grid.major = element_blank(),
		panel.grid.minor = element_blank(),
	)

ggsave("../30_diversity-analysis/plots/beta-diversity/jc_gender.png", units = "px", width = w, height = h, create.dir = TRUE)
	
```

#### Human
```{r}
human_jc_dm <-phyloseq::distance(human, method = "jaccard", binary = TRUE)
adonis2(human_jc_dm ~ trt*gender, data = data.frame(humandf))
```
Community composition is not significantly different between treatments or gender.

```{r}
human_pcoa_jc <- ordinate(human, method="PCoA", distance=human_jc_dm)

plot_ordination(human, human_pcoa_jc, color = "trt") +
	#stat_ellipse(alpha = 0.5) +
	scale_color_manual(values = pal) +
  labs(col = "Treatment") +
	theme_bw(base_size = font_size) +
	theme(
		panel.grid.major = element_blank(),
		panel.grid.minor = element_blank(),
	)

ggsave("../30_diversity-analysis/plots/beta-diversity/jc_trt_human.png", units = "px", width = w, height = h, create.dir = TRUE)

plot_ordination(human, human_pcoa_jc, color = "gender") +
	#stat_ellipse(alpha = 0.5) +
	scale_color_manual(values = pal3) +
  labs(col = "Gender") +
	theme_bw(base_size = font_size) +
	theme(
		panel.grid.major = element_blank(),
		panel.grid.minor = element_blank(),
	)

ggsave("../30_diversity-analysis/plots/beta-diversity/jc_gender_human.png", units = "px", width = w, height = h, create.dir = TRUE)
```

#### Mouse
```{r}
mouse_jc_dm <-phyloseq::distance(mouse, method = "jaccard", binary = TRUE)
adonis2(mouse_jc_dm ~ gender*trt*starch_type, data = data.frame(mousedf))
pw <- pairwiseAdonis::pairwise.adonis2(mouse_jc_dm ~ starch_type, data = data.frame(mousedf))

print("significantly different pairs: ")
for (i in range(2,(length(pw)))){
	pair <- pw[[i]]
	if (pair$`Pr(>F)`[1] < 0.05) {
		print(names(pw)[i])
	}
}
```
Community composition is significantly different between genders, treatments, starch types, at least 2 gender:trt pairs and at least 2 gender:starch_type pairs. The significantly different starch_type groups are PTB-BEP and CKP-CTL. 

```{r}
mouse_pcoa_jc <- ordinate(mouse, method="PCoA", distance=mouse_jc_dm)

plot_ordination(mouse, mouse_pcoa_jc, color = "trt") +
	#stat_ellipse(alpha = 0.5) +
	scale_color_manual(values = pal[3:4]) +
  labs(col = "Treatment") +
	theme_bw(base_size = font_size) +
	theme(
		panel.grid.major = element_blank(),
		panel.grid.minor = element_blank(),
	)

ggsave("../30_diversity-analysis/plots/beta-diversity/jc_trt_mouse.png", units = "px", width = w, height = h, create.dir = TRUE)

plot_ordination(mouse, mouse_pcoa_jc, color = "starch_type") +
	#stat_ellipse(alpha = 0.5) +
	scale_color_manual(values = pal2) +
  labs(col = "Starch Type") +
	theme_bw(base_size = font_size) +
	theme(
		panel.grid.major = element_blank(),
		panel.grid.minor = element_blank(),
	)

ggsave("../30_diversity-analysis/plots/beta-diversity/jc_type_mouse.png", units = "px", width = w, height = h, create.dir = TRUE)

plot_ordination(mouse, mouse_pcoa_jc, color = "gender") +
	#stat_ellipse(alpha = 0.5) +
	scale_color_manual(values = pal3) +
  labs(col = "Gender") +
	theme_bw(base_size = font_size) +
	theme(
		panel.grid.major = element_blank(),
		panel.grid.minor = element_blank(),
	)

ggsave("../30_diversity-analysis/plots/beta-diversity/jc_gender_mouse.png", units = "px", width = w, height = h, create.dir = TRUE)
```

### Unweighted Unifrac

#### Merged
```{r}
uu_dm <- phyloseq::distance(starch, method = "unifrac")
adonis2(uu_dm ~ group*gender, data = data.frame(starchdf))
adonis2(uu_dm ~ cohort*trt*gender, data = data.frame(starchdf))
```
There is a significant difference in community composition and phylogeny between cohorts, genders and at least 2 group:gender pairs.

```{r}
pcoa_uu <- ordinate(starch, method="PCoA", distance=uu_dm)

plot_ordination(starch, pcoa_uu, color = "group") +
	#stat_ellipse(alpha = 0.5) +
	scale_color_manual(values = pal) +
  labs(col = "Group") +
	theme_bw(base_size = font_size) +
	theme(
		panel.grid.major = element_blank(),
		panel.grid.minor = element_blank(),
	)

ggsave("../30_diversity-analysis/plots/beta-diversity/uu_group.png", units = "px", width = w, height = h, create.dir = TRUE)

plot_ordination(starch, pcoa_uu, color = "gender") +
	#stat_ellipse(alpha = 0.5) +
	scale_color_manual(values = pal3) +
  labs(col = "Gender") +
	theme_bw(base_size = font_size) +
	theme(
		panel.grid.major = element_blank(),
		panel.grid.minor = element_blank(),
	)

ggsave("../30_diversity-analysis/plots/beta-diversity/uu_gender.png", units = "px", width = w, height = h, create.dir = TRUE)
	
```

#### Human
```{r}
human_uu_dm <-phyloseq::distance(human, method = "unifrac")
adonis2(human_uu_dm ~ trt*gender, data = data.frame(humandf), permutations = 9999)
```
There is a significant difference in community composition and phylogeny between genders. 

```{r}
human_pcoa_uu <- ordinate(human, method="PCoA", distance=human_uu_dm)

plot_ordination(human, human_pcoa_uu, color = "trt") +
	#stat_ellipse(alpha = 0.5) +
	scale_color_manual(values = pal) +
  labs(col = "Treatment") +
	theme_bw(base_size = font_size) +
	theme(
		panel.grid.major = element_blank(),
		panel.grid.minor = element_blank(),
	)

ggsave("../30_diversity-analysis/plots/beta-diversity/uu_trt_human.png", units = "px", width = w, height = h, create.dir = TRUE)

plot_ordination(human, human_pcoa_uu, color = "gender") +
	#stat_ellipse(alpha = 0.5) +
	scale_color_manual(values = pal3) +
  labs(col = "Gender") +
	theme_bw(base_size = font_size) +
	theme(
		panel.grid.major = element_blank(),
		panel.grid.minor = element_blank(),
	)

ggsave("../30_diversity-analysis/plots/beta-diversity/uu_gender_human.png", units = "px", width = w, height = h, create.dir = TRUE)
```

#### Mouse
```{r}
mouse_uu_dm <-phyloseq::distance(mouse, method = "unifrac")
adonis2(mouse_uu_dm ~ trt*gender*starch_type, data = data.frame(mousedf))
```
There no significant difference in community composition and phylogeny between treatments but there is a difference between genders and at least 2 gender:starch_type pairs. 

```{r}
mouse_pcoa_uu <- ordinate(mouse, method="PCoA", distance=mouse_uu_dm)

plot_ordination(mouse, mouse_pcoa_uu, color = "trt") +
	#stat_ellipse(alpha = 0.5) +
	scale_color_manual(values = pal[3:4]) +
  labs(col = "Treatment") +
	theme_bw(base_size = font_size) +
	theme(
		panel.grid.major = element_blank(),
		panel.grid.minor = element_blank(),
	)

ggsave("../30_diversity-analysis/plots/beta-diversity/uu_trt_mouse.png", units = "px", width = w, height = h, create.dir = TRUE)

plot_ordination(mouse, mouse_pcoa_uu, color = "starch_type") +
	#stat_ellipse(alpha = 0.5) +
	scale_color_manual(values = pal2) +
  labs(col = "Starch Type") +
	theme_bw(base_size = font_size) +
	theme(
		panel.grid.major = element_blank(),
		panel.grid.minor = element_blank(),
	)

ggsave("../30_diversity-analysis/plots/beta-diversity/uu_type_mouse.png", units = "px", width = w, height = h, create.dir = TRUE)

plot_ordination(mouse, mouse_pcoa_uu, color = "gender") +
	#stat_ellipse(alpha = 0.5) +
	scale_color_manual(values = pal3) +
  labs(col = "Gender") +
	theme_bw(base_size = font_size) +
	theme(
		panel.grid.major = element_blank(),
		panel.grid.minor = element_blank(),
	)

ggsave("../30_diversity-analysis/plots/beta-diversity/uu_gender_mouse.png", units = "px", width = w, height = h, create.dir = TRUE)
```
there's an interesting clustering, but we'd need more metadata to see what is causing this difference. 

### Weighted Unifrac
#### Merged
```{r}
wu_dm <- phyloseq::distance(starch, method = "wunifrac")
adonis2(wu_dm ~ group*gender, data = data.frame(starchdf), permutations = 9999)
adonis2(wu_dm ~ cohort*trt*gender, data = data.frame(starchdf), permutations = 9999)
```
Significant difference in community structure and phylogeny between cohorts, genders and at least 2 trt:gender pairs.  

```{r}
pcoa_wu <- ordinate(starch, method="PCoA", distance=wu_dm)

plot_ordination(starch, pcoa_wu, color = "group") +
	#stat_ellipse(alpha = 0.5) +
	scale_color_manual(values = pal) +
  labs(col = "Group") +
	theme_bw(base_size = font_size) +
	theme(
		panel.grid.major = element_blank(),
		panel.grid.minor = element_blank(),
	)

ggsave("../30_diversity-analysis/plots/beta-diversity/wu_group.png", units = "px", width = w, height = h, create.dir = TRUE)

plot_ordination(starch, pcoa_wu, color = "gender") +
	#stat_ellipse(alpha = 0.5) +
	scale_color_manual(values = pal3) +
  labs(col = "Gender") +
	theme_bw(base_size = font_size) +
	theme(
		panel.grid.major = element_blank(),
		panel.grid.minor = element_blank(),
	)

ggsave("../30_diversity-analysis/plots/beta-diversity/wu_gender.png", units = "px", width = w, height = h, create.dir = TRUE)
	
```

#### Human
```{r}
human_wu_dm <-phyloseq::distance(human, method = "wunifrac")
adonis2(human_wu_dm ~ trt*gender, data = data.frame(humandf), permutations = 9999)
```
There is no significant difference in community structure and phylogeny between treatments, genders or the interaction between them. 

```{r}
human_pcoa_wu <- ordinate(human, method="PCoA", distance=human_wu_dm)

plot_ordination(human, human_pcoa_wu, color = "trt") +
	#stat_ellipse(alpha = 0.5) +
	scale_color_manual(values = pal[1:2]) +
  labs(col = "Treatment") +
	theme_bw(base_size = font_size) +
	theme(
		panel.grid.major = element_blank(),
		panel.grid.minor = element_blank(),
	)

ggsave("../30_diversity-analysis/plots/beta-diversity/wu_trt_human.png", units = "px", width = w, height = h, create.dir = TRUE)

plot_ordination(human, human_pcoa_wu, color = "gender") +
	#stat_ellipse(alpha = 0.5) +
	scale_color_manual(values = pal3) +
  labs(col = "Gender") +
	theme_bw(base_size = font_size) +
	theme(
		panel.grid.major = element_blank(),
		panel.grid.minor = element_blank(),
	)

ggsave("../30_diversity-analysis/plots/beta-diversity/wu_gender_human.png", units = "px", width = w, height = h, create.dir = TRUE)
```

#### Mouse
```{r}
mouse_wu_dm <-phyloseq::distance(mouse, method = "wunifrac")
adonis2(mouse_wu_dm ~ trt*gender*starch_type, data = data.frame(mousedf))
pw <- pairwiseAdonis::pairwise.adonis2(mouse_wu_dm ~ starch_type, data = data.frame(mousedf))

print("significantly different pairs: ")
for (i in range(2,(length(pw)))){
	pair <- pw[[i]]
	if (pair$`Pr(>F)`[1] < 0.05) {
		print(names(pw)[i])
	}
}
```
There is a significant difference in community structure and phylogeny between treatments, genders, at least 2 starch types, at least 2 trt:gender pairs, and at least 2 gender:starch_type pairs. The significantly different starch type pairs are: PTB-BEP and CKP-CTL. 

```{r}
mouse_pcoa_wu <- ordinate(mouse, method="PCoA", distance=mouse_wu_dm)

plot_ordination(mouse, mouse_pcoa_wu, color = "trt") +
	#stat_ellipse(alpha = 0.5) +
	scale_color_manual(values = pal[3:4]) +
  labs(col = "Treatment") +
	theme_bw(base_size = font_size) +
	theme(
		panel.grid.major = element_blank(),
		panel.grid.minor = element_blank(),
	)

ggsave("../30_diversity-analysis/plots/beta-diversity/wu_trt_mouse.png", units = "px", width = w, height = h, create.dir = TRUE)

plot_ordination(mouse, mouse_pcoa_wu, color = "starch_type") +
	#stat_ellipse(alpha = 0.5) +
	scale_color_manual(values = pal2) +
  labs(col = "Starch Type") +
	theme_bw(base_size = font_size) +
	theme(
		panel.grid.major = element_blank(),
		panel.grid.minor = element_blank(),
	)

ggsave("../30_diversity-analysis/plots/beta-diversity/wu_type_mouse.png", units = "px", width = w, height = h, create.dir = TRUE)

plot_ordination(mouse, mouse_pcoa_wu, color = "gender") +
	#stat_ellipse(alpha = 0.5) +
	scale_color_manual(values = pal3) +
  labs(col = "Gender") +
	theme_bw(base_size = font_size) +
	theme(
		panel.grid.major = element_blank(),
		panel.grid.minor = element_blank(),
	)

ggsave("../30_diversity-analysis/plots/beta-diversity/wu_gender_mouse.png", units = "px", width = w, height = h, create.dir = TRUE)
```
